refactor: mapper 위치 변경
