+++ data/src/main/java/com/housweet/data/repository/AppSettingRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.housweet.data.repository
 
 import com.housweet.data.network.AppSettingRemoteDataSource
-import com.housweet.data.network.dto.toDomain
+import com.housweet.data.mapper.toDomain
 import com.housweet.domain.model.AppSettingCategory
 import com.housweet.domain.model.AppSettingItem
 import com.housweet.domain.repository.AppSettingRepository
